requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
gunicorn==22.0.0
sqlalchemy==2.0.31
pymysql==1.1.1
cryptography==42.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-multipart==0.0.9
bleach==6.1.0
